requirements.txt
openai>=1.0.0
python-docx>=1.0.0
pdfplumber>=0.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.60.0
